+++ src/scriptlets/scriptlets-list.ts
@@ -78,6 +78,7 @@ export { trustedReplaceArgument } from './trusted-replace-argument';
 export { preventInnerHTML } from './prevent-innerHTML';
 export { preventConstructor } from './prevent-constructor';
 export { removeRequestQueryParameter } from './remove-request-query-parameter';
+export { preventNavigation } from './prevent-navigation';
 // redirects as scriptlets
 // https://github.com/AdguardTeam/Scriptlets/issues/300
 export { AmazonApstag } from './amazon-apstag';