+++ tests/scriptlets/index.test.js
@@ -61,3 +61,4 @@ import './trusted-suppress-native-method.test';
 import './prevent-canvas.test';
 import './trusted-replace-argument.test';
 import './remove-request-query-parameter.test';
+import './prevent-navigation.test';